mongolog: rename, comment out intel-specific code
