+++ CMakeLists.txt
@@ -10,6 +10,8 @@ include($ENV{ROS_ROOT}/core/rosbuild/rosbuild.cmake)
 #set(ROS_BUILD_TYPE RelWithDebInfo)
 
 rosbuild_init()
+rosbuild_add_boost_directories()
+
 
 set(EXECUTABLE_OUTPUT_PATH ${PROJECT_SOURCE_DIR}/bin)
 set(LIBRARY_OUTPUT_PATH ${PROJECT_SOURCE_DIR}/lib)
@@ -18,7 +20,12 @@ set(LIBRARY_OUTPUT_PATH ${PROJECT_SOURCE_DIR}/lib)
 #rosbuild_gensrv()
 
 link_libraries(mongoclient)
-rosbuild_add_executable(rosmongolog_tf src/rosmongolog_tf.cpp)
-rosbuild_add_executable(rosmongolog_pcl src/rosmongolog_pcl.cpp)
-rosbuild_add_executable(rosmongolog_cimg src/rosmongolog_cimg.cpp)
-rosbuild_add_executable(rosmongolog_trimesh src/rosmongolog_trimesh.cpp)
+rosbuild_add_executable(mongolog_tf src/mongolog_tf.cpp)
+rosbuild_add_executable(mongolog_pcl src/mongolog_pcl.cpp)
+rosbuild_add_executable(mongolog_cimg src/mongolog_cimg.cpp)
+# rosbuild_add_executable(rosmongolog_trimesh src/rosmongolog_trimesh.cpp)
+rosbuild_link_boost(mongolog_cimg system)
+rosbuild_link_boost(mongolog_pcl system)
+rosbuild_link_boost(mongolog_tf system)
+#rosbuild_link_boost(mongolog_trimesh system)
+
