+++ bin/mongolog
@@ -1,0 +1,1 @@
+../src/mongolog.py
