--- bin/rosmongolog_mp
@@ -1,1 +1,0 @@
-../src/rosmongolog_mp.py
